Entity Named Notes Section
#2324
