+++ waltz-ng/client/actor/actor-view.html
@@ -20,6 +20,12 @@
 
     <br>
 
+
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityRef">
+    </waltz-entity-named-notes-section>
+
+
     <!-- APPS -->
     <waltz-apps-section parent-entity-ref="ctrl.entityRef"
                         scope="EXACT">
