+++ waltz-ng/client/actor/components/actor-overview.html
@@ -23,10 +23,4 @@
         </div>
     </div>
 
-    <div class="row">
-        <div class="col-sm-12">
-            <waltz-entity-named-notes-panel parent-entity-ref="$ctrl.entityRef">
-            </waltz-entity-named-notes-panel>
-        </div>
-    </div>
 </div>
