+++ waltz-ng/client/applications/app-view.html
@@ -24,6 +24,11 @@
 
     <br>
 
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityRef">
+    </waltz-entity-named-notes-section>
+
+
     <!-- RATINGS -->
     <waltz-measurable-rating-app-section parent-entity-ref="ctrl.entityRef"
                                          application='ctrl.app'
