+++ waltz-ng/client/applications/components/app-overview.html
@@ -149,11 +149,6 @@
         </div>
     </div>
 
-    <br>
-    <waltz-entity-named-notes-panel parent-entity-ref="$ctrl.entityRef">
-    </waltz-entity-named-notes-panel>
-
-
 
     <!--  TEMPLATES -->
