+++ waltz-ng/client/change-initiative/change-initiative-view.html
@@ -79,13 +79,15 @@
                         </div>
                     </div>
 
-                    <br>
-                    <waltz-entity-named-notes-panel parent-entity-ref="ctrl.entityRef">
-                    </waltz-entity-named-notes-panel>
                 </div>
             </div>
         </div>
     </div>
+
+
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityRef">
+    </waltz-entity-named-notes-section>
 
 
     <!-- BOOKMARKS -->
