+++ waltz-ng/client/data-types/data-type-view.html
@@ -21,6 +21,11 @@
     </waltz-data-type-overview>
 
 
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityRef">
+    </waltz-entity-named-notes-section>
+
+
     <!-- AUTH SOURCES -->
     <waltz-section name="Authoritative Sources"
                    id="authoritative-sources-section"
