--- waltz-ng/client/entity-named-note/components/entity-named-notes-panel.scss
@@ -1,8 +1,0 @@
-.waltz-entity-named-notes {
-    padding-left: 20px;
-    padding-right: 20px;
-
-    .wenn-table {
-        border-left: 1px solid #0000FF;
-    }
-}
