+++ waltz-ng/client/entity-named-note/components/panel/entity-named-notes-panel.scss
@@ -1,0 +1,2 @@
+.waltz-entity-named-notes {
+}
