+++ waltz-ng/client/entity-named-note/components/section/entity-named-notes-section.html
@@ -1,0 +1,48 @@
+<!--
+  ~ /*
+  ~  * Waltz - Enterprise Architecture
+  ~  * Copyright (C) 2017  Khartec Ltd.
+  ~  *
+  ~  * This program is free software: you can redistribute it and/or modify
+  ~  * it under the terms of the GNU Lesser General Public License as published by
+  ~  * the Free Software Foundation, either version 3 of the License, or
+  ~  * (at your option) any later version.
+  ~  *
+  ~  * This program is distributed in the hope that it will be useful,
+  ~  * but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~  * GNU Lesser General Public License for more details.
+  ~  *
+  ~  * You should have received a copy of the GNU Lesser General Public License
+  ~  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  ~  */
+  -->
+
+<div ng-show="$ctrl.showSection()">
+    <waltz-section name="Additional Notes"
+                   icon="sticky-note-o"
+                   id="entity-notes-section">
+
+        <waltz-section-actions>
+            <div waltz-has-role-for-entity-kind="{{ $ctrl.parentEntityRef.kind }}">
+                <button type="button"
+                        class="btn btn-xs btn-primary"
+                        ng-if="$ctrl.hasCreatableNoteTypes && ! $ctrl.creatingNote"
+                        ng-click="$ctrl.showEditor()">
+                    Add Note
+                </button>
+            </div>
+        </waltz-section-actions>
+
+
+        <waltz-entity-named-notes-panel parent-entity-ref="$ctrl.parentEntityRef"
+                                        notes="$ctrl.notes"
+                                        all-note-types="$ctrl.allNoteTypes"
+                                        on-save="$ctrl.saveNote"
+                                        on-delete="$ctrl.deleteNote"
+                                        creating-note="$ctrl.creatingNote"
+                                        on-editor-dismiss="$ctrl.editorDismiss">
+        </waltz-entity-named-notes-panel>
+
+    </waltz-section>
+</div>
