+++ waltz-ng/client/measurable/components/summary/measurable-summary.html
@@ -6,8 +6,6 @@
         <div class="col-sm-10 text-muted">
             <waltz-markdown text="$ctrl.measurable.description">
             </waltz-markdown>
-            <waltz-entity-named-notes-panel parent-entity-ref="$ctrl.entityRef">
-            </waltz-entity-named-notes-panel>
         </div>
     </div>
