+++ waltz-ng/client/measurable/measurable-view.html
@@ -28,6 +28,10 @@
 
     <br>
 
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityReference">
+    </waltz-entity-named-notes-section>
+
 
     <!-- RATINGS -->
     <waltz-measurable-rating-explorer-section parent-entity-ref="ctrl.entityReference"
