+++ waltz-ng/client/org-units/components/overview/org-unit-overview.html
@@ -12,8 +12,6 @@
                 <br>
             </span>
 
-            <waltz-entity-named-notes-panel parent-entity-ref="$ctrl.entityRef">
-            </waltz-entity-named-notes-panel>
             <hr/>
 
             <div class="container-fluid">
