+++ waltz-ng/client/org-units/unit-view.html
@@ -27,6 +27,11 @@
     </waltz-org-unit-overview>
 
 
+    <!--Entity Named Notes-->
+    <waltz-entity-named-notes-section parent-entity-ref="ctrl.entityRef">
+    </waltz-entity-named-notes-section>
+
+
     <!-- RATINGS -->
     <waltz-measurable-ratings-browser-section id="ratings-explorer-section"
                                                parent-entity-ref="ctrl.entityRef"
