+++ waltz-ng/style/_waltz.scss
@@ -24,7 +24,7 @@
 @import "../client/complexity/components/complexity-bar-chart";
 @import "../client/data-type-usage/components/editor/app-data-type-usage-editor";
 @import "../client/entity/components/immediate-hierarchy-navigator/immediate-hierarchy-navigator";
-@import "../client/entity-named-note/components/entity-named-notes-panel";
+@import "../client/entity-named-note/components/panel/entity-named-notes-panel";
 @import "../client/entity/components/entity-selector/entity-selector";
 @import "../client/entity-statistics/directives/entity-statistic";
 @import "../client/entity-statistics/components/entity-statistic-summary-card";
